Tests: ??? Fix failing test
For some reason, awaiting the `Task.Delay` call here makes the continuation
return a `None<double>`, with no visible exception thrown or anything.

Awaiting `Task.FromResult` seems to work fine, though, and will do for test
purposes. If it keeps happening in the real world, more research will need to
be done to figure out what's causing it. Maybe some weird interaction with
the OptionAsync pieces?

If this